Split code into different modules
