+++ submissions/week-2/day-3/Favour-Adedeji/task/Cargo.toml
@@ -3,4 +3,8 @@ name = "task"
 version = "0.1.0"
 edition = "2024"
 
+[lib]
+name = "web3bridge"
+path = "src/lib.rs"
+
 [dependencies]
