+++ submissions/week-2/day-3/Favour-Adedeji/task/src/enums.rs
@@ -1,0 +1,15 @@
+#[derive(Debug)]
+pub enum Department {
+  MEDIA,
+  IT,
+  MANAGER,
+  SOCIAL_MEDIA,
+  TECHNICIAN_SUPERVISOR,
+  KITCHEN_STAFF,
+}
+
+#[derive(Debug, PartialEq)]
+pub enum Status {
+  ACTIVE,
+  TERMINATED,
+}
