+++ submissions/week-2/day-3/Favour-Adedeji/task/src/state.rs
@@ -1,0 +1,7 @@
+use crate::enums::{Department, Status};
+
+pub struct EmployeeDetails {
+    pub name: String,
+    pub department: Department,
+    pub is_terminated: Status,
+}
